+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add operating mode support to clean mode
+
 ## [2.0.2] - 2026-05-20
 
 - Move @matter/types to as deps